pyproject.toml
[project]
name = "agent-context-mpp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "cryptography>=43.0.0",
    "eth-account",
    "eth-utils",
    "fastapi",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings",
    "pympp[tempo]",
    "python-frontmatter",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
    "web3>=7.0.0",
]

[dependency-groups]
dev = [
    "pyright",
    "pytest",
    "ruff",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.format]
line-ending = "lf"

[tool.pyright]
include = ["app", "tests"]
pythonVersion = "3.12"
typeCheckingMode = "strict"

[tool.pytest.ini_options]
pythonpath = ["."]
